+++ v3.x.x/main/autopatcher/AssetBundleMgr.cs
