+++ v3.x.x/main/autopatcher/ConfigMgr.cs
