+++ v3.x.x/main/autopatcher/LuaMgr.cs
