+++ v3.x.x/main/autopatcher/PathMgr.cs
