+++ v3.x.x/main/autopatcher/Program.cs
@@ -26,7 +26,8 @@ namespace Azurlane
         internal static bool Abort;
         internal static List<string> ListOfLua;
         internal static Dictionary<Mods, bool> ListOfMod;
-        internal static string DirName = "CAB-android64";
+        internal static string DirName = "CAB-android";
+        internal static string arch; 
 
         private static List<Action> _listOfAction;
 
@@ -120,7 +121,7 @@ namespace Azurlane
                         Utils.Write("[Obsolete Autopatcher version]", true, true);
                         Utils.Write("Download the latest version from:", true, true);
                         Utils.Write(Properties.Resources.Repository, true, true);
-                        Abort = true;
+                        Abort = false;
                     }
                 }
             }
@@ -139,7 +140,7 @@ namespace Azurlane
 
                 foreach (var mod in ListOfMod.Keys)
                 {
-                    var modName = "scripts64-" + mod.ToString().ToLower().Replace("_", "-");
+                    var modName = $"scripts32-{mod.ToString().ToLower().Replace("_", "-")}";
                     if (File.Exists(PathMgr.Temp(modName))) File.Delete(PathMgr.Temp(modName));
                     if (Directory.Exists(PathMgr.Lua(modName).Replace($"\\{DirName}", ""))) Utils.Rmdir(PathMgr.Lua(modName).Replace($"\\{DirName}", ""));
                 }
@@ -166,7 +167,7 @@ namespace Azurlane
             ConfigMgr.Initialize();
 
             Message();
-            CheckVersion();
+            //CheckVersion();
             CheckDependencies();
 
             AddLua(Properties.Resources.Aircraft);
@@ -228,6 +229,19 @@ namespace Azurlane
                 goto END;
             }
 
+            if (args.Contains("64"))
+            {
+                arch = @"64";
+                Utils.Write(@"Selected scripts is 64 bits",true,true);
+            }
+            else if (args.Contains("32"))
+            {
+                arch = @"32";
+                Utils.Write(@"Selected scripts is 32 bits",true,true);
+            }
+
+            DirName += arch;
+            
             var filePath = Path.GetFullPath(args[0]);
             var fileDirectoryPath = Path.GetDirectoryName(filePath);
             var fileName = Path.GetFileName(filePath);
@@ -272,7 +286,7 @@ namespace Azurlane
                     {
                         try {
                             Utils.LogInfo("Decrypting AssetBundle...", true, false);
-                            Utils.Command($"Azcli64.exe --dev --decrypt \"{PathMgr.Temp(fileName)}\"");
+                            Utils.Command($"Azcli.exe --dev --decrypt \"{PathMgr.Temp(fileName)}\"");
                             Utils.Write(" <done>", false, true);
                         }
                         catch (Exception e)
@@ -285,7 +299,7 @@ namespace Azurlane
                     {
                         try {
                             Utils.LogInfo("Unpacking AssetBundle...", true, false);
-                            Utils.Command($"Azcli64.exe --dev --unpack \"{PathMgr.Temp(fileName)}\"");
+                            Utils.Command($"Azcli.exe --dev --unpack \"{PathMgr.Temp(fileName)}\"");
                             Utils.Write(" <done>", false, true);
                         }
                         catch (Exception e)
@@ -300,7 +314,7 @@ namespace Azurlane
                             var showDoneMessage = true;
                             Utils.LogInfo("Decrypting Lua...", true, false);
                             foreach (var lua in ListOfLua) {
-                                Utils.Command($"Azcli64.exe --dev --unlock \"{PathMgr.Lua(fileName, lua)}\"");
+                                Utils.Command($"Azcli.exe --dev --unlock \"{PathMgr.Lua(fileName, lua)}\"");
 
                                 if (LuaMgr.CheckLuaState(PathMgr.Lua(fileName, lua)) != LuaMgr.State.Encrypted)
                                     break;
@@ -326,7 +340,7 @@ namespace Azurlane
                             foreach (var lua in ListOfLua) {
                                 tasks.Add(Task.Factory.StartNew(() =>
                                 {
-                                    Utils.Command($"Azcli64.exe --dev --decompile \"{PathMgr.Lua(fileName, lua)}\"");
+                                    Utils.Command($"Azcli.exe --dev --decompile \"{PathMgr.Lua(fileName, lua)}\"");
                                     Utils.Write($@" {index}/{ListOfLua.Count}", false, false);
                                     index++;
                                 }));
@@ -348,7 +362,7 @@ namespace Azurlane
                             {
                                 if (mod.Value)
                                 {
-                                    var modName = ("scripts64-" + mod.Key).ToLower().Replace("_", "-");
+                                    var modName = ("scripts32-" + mod.Key).ToLower().Replace("_", "-");
 
                                     if (!Directory.Exists(PathMgr.Lua(modName)))
                                         Directory.CreateDirectory(PathMgr.Lua(modName));
@@ -412,11 +426,11 @@ namespace Azurlane
                             {
                                 if (mod.Value)
                                 {
-                                    var modName = ("scripts64-" + mod.Key).ToLower().Replace("_", "-");
+                                    var modName = ("scripts32-" + mod.Key).ToLower().Replace("_", "-");
                                     foreach (var lua in ListOfLua) {
                                         tasks.Add(Task.Factory.StartNew(() =>
                                         {
-                                            Utils.Command($"Azcli64.exe --dev --recompile \"{PathMgr.Lua(modName,lua)}\"");
+                                            Utils.Command($"Azcli.exe --dev --recompile \"{PathMgr.Lua(modName,lua)}\"");
                                         }));
                                     }
                                 }
@@ -439,10 +453,10 @@ namespace Azurlane
                             {
                                 if (mod.Value)
                                 {
-                                    var modName = ("scripts64-" + mod.Key).ToLower().Replace("_", "-");
+                                    var modName = ("scripts32-" + mod.Key).ToLower().Replace("_", "-");
 
                                     foreach (var lua in ListOfLua) {
-                                        Utils.Command($"Azcli64.exe --dev --lock \"{PathMgr.Lua(modName, lua)}\"");
+                                        Utils.Command($"Azcli.exe --dev --lock \"{PathMgr.Lua(modName, lua)}\"");
 
                                         if (LuaMgr.CheckLuaState(PathMgr.Lua(modName, lua)) != LuaMgr.State.Decrypted)
                                             break;
@@ -472,11 +486,11 @@ namespace Azurlane
                             {
                                 if (mod.Value)
                                 {
-                                    var modName = ("scripts64-" + mod.Key).ToLower().Replace("_", "-");
+                                    var modName = ("scripts32-" + mod.Key).ToLower().Replace("_", "-");
 
                                     tasks.Add(Task.Factory.StartNew(() =>
                                     {
-                                        Utils.Command($"Azcli64.exe --dev --repack \"{PathMgr.Temp(modName)}\"");
+                                        Utils.Command($"Azcli.exe --dev --repack \"{PathMgr.Temp(modName)}\"");
                                         Utils.Write($@" {index}/{ListOfMod.Count(x => x.Value)}", false, false);
                                         index++;
                                     }));
@@ -499,8 +513,8 @@ namespace Azurlane
                             {
                                 if (mod.Value)
                                 {
-                                    var modName = ("scripts64-" + mod.Key).ToLower().Replace("_", "-");
-                                    Utils.Command($"Azcli64.exe --dev --encrypt \"{PathMgr.Temp(modName)}\"");
+                                    var modName = ("scripts32-" + mod.Key).ToLower().Replace("_", "-");
+                                    Utils.Command($"Azcli.exe --dev --encrypt \"{PathMgr.Temp(modName)}\"");
                                 }
                             }
                             Utils.Write(" <done>", false, true);
@@ -519,7 +533,7 @@ namespace Azurlane
                             {
                                 if (mod.Value)
                                 {
-                                    var modName = ("scripts64-" + mod.Key).ToLower().Replace("_", "-");
+                                    var modName = ("scripts32-" + mod.Key).ToLower().Replace("_", "-");
 
                                     if (File.Exists(Path.Combine(fileDirectoryPath, modName)))
                                         File.Delete(Path.Combine(fileDirectoryPath, modName));
@@ -591,7 +605,7 @@ namespace Azurlane
                 {
                     if (keyValuePair.Value)
                     {
-                        string name = "scripts64-" + keyValuePair.Key.ToString().ToLower().Replace("_", "-");
+                        string name = $"scripts32-{keyValuePair.Key.ToString().ToLower().Replace("_", "-")}";
                         string path2 = keyValuePair.Key.ToString().ToLower().Replace("_", "-");
                         Directory.CreateDirectory(Path.Combine(text2, path2));
                         foreach (string text3 in Directory.GetFiles(PathMgr.Assets(name), "*.*", SearchOption.AllDirectories))
