+++ v3.x.x/main/autopatcher/Properties/AssemblyInfo.cs
