+++ v3.x.x/main/autopatcher/Properties/Resources.resx
