+++ v3.x.x/main/autopatcher/Resources/Configuration.ini
