+++ v3.x.x/main/autopatcher/Utils.cs
