--- v3.x.x/main/autopatcher32/Resources/Configuration.ini
@@ -1,59 +1,0 @@
-[Common]
-Version=v3.0.1
-DevelopmentMode=true
-
-[Path]
-Tmp=tmp
-Thirdparty=3rdparty
-
-[Mods]
-GodMode=false
-WeakEnemy=false
-GodMode+Damage=false
-GodMode+Cooldown=false
-GodMode+WeakEnemy=false
-GodMode+Damage+Cooldown=false
-GodMode+Damage+WeakEnemy=true
-GodMode+Damage+Cooldown+WeakEnemy=false
-
-[Aircraft]
-Hp=ignore
-HpGrowth=ignore
-Accuracy=0
-AccuracyGrowth=0
-AttackPower=0
-AttackPowerGrowth=0
-CrashDamage=0
-Speed=ignore
-
-[Weapon]
-Damage=300
-ReloadMax=120
-
-[Enemy]
-AntiAir=ignore
-AntiAirGrowth=ignore
-AntiSubmarine=ignore
-Armor=ignore
-ArmorGrowth=ignore
-Cannon=ignore
-CannonGrowth=ignore
-Evasion=ignore
-EvasionGrowth=ignore
-Hit=ignore
-HitGrowth=ignore
-Hp=50
-HpGrowth=100
-Luck=ignore
-LuckGrowth=ignore
-Reload=1
-ReloadGrowth=1
-Speed=15
-SpeedGrowth=17
-Torpedo=1
-TorpedoGrowth=1
-RemoveSkill=true
-
-[Other]
-ReplaceSkin=true
-EasyMode=true
