--- v3.x.x/main/autopatcher32/status.txt
@@ -1,1 +1,0 @@
-ok
