--- v3.x.x/main/autopatcher32/version.txt
@@ -1,1 +1,0 @@
-v3.0.1
