--- v3.x.x/main/autopatcher64/status.txt
@@ -1,1 +1,0 @@
-ok
